requirements.txt
folium 
plotly 
pandas 
numpy 
matplotlib 
seaborn
rich
pyyaml
gitpython
scipy
hydra-core
omegaconf
mlflow
ray